+++ rope/base/oi/type_hinting/providers/numpydocstrings.py
@@ -3,6 +3,7 @@ Some code extracted (or based on code) from:
 https://github.com/davidhalter/jedi/blob/b489019f5bd5750051122b94cc767df47751ecb7/jedi/evaluate/docstrings.py
 Thanks to @davidhalter for this utils under MIT License.
 """
+
 import re
 
 from rope.base.ast import literal_eval