+++ rope/base/prefs.py
@@ -18,6 +18,21 @@ class AutoimportPrefs:
         default=False, description="Cache underlined (private) modules")
     memory: bool = field(default=None, description="Cache in memory instead of disk")
     parallel: bool = field(default=True, description="Use multiple processes to parse")
+    aliases: List[Tuple[str, str]] = field(
+        default_factory=lambda : [
+            ("np", "numpy"),
+            ("pd", "pandas"),
+            ("plt", "matplotlib.pyplot"),
+            ("sns", "seaborn"),
+            ("tf", "tensorflow"),
+            ("sk", "sklearn"),
+            ("sm", "statsmodels"),
+        ],
+        description=dedent("""
+            Aliases for module names.  For example, `[('np', 'numpy')]` makes rope recommend
+            ``import numpy as np``.
+        """),
+    )
 
 
 @dataclass
@@ -147,7 +162,6 @@ Builtin and c-extension modules that are allowed to be imported and inspected by
             appear in the importing namespace.
         """),
     )
-
     prefer_module_from_imports: bool = field(
         default=False,
         description=dedent("""