+++ rope/refactor/__init__.py
@@ -45,6 +45,7 @@ TaskHandle` parameter.  You can use this object for stopping or
 monitoring the progress of refactorings.
 
 """
+
 from rope.refactor.importutils import ImportOrganizer  # noqa
 from rope.refactor.topackage import ModuleToPackage  # noqa